feat: sending a POST without required fields returns "required field(s) missing"
Wrote a test to make sure it works, because I couldnt send a POST request via the web interface,
because HTML-wise it already did not allow for the input fields to be empty.
So writing a test and forcing a POST request with missing fields seemed like a good option and it worked.