feat: return full object on POST, allow for filter queries and return filtered array
Theres not more to explain. Only thing I am worried about is that the FCC test suite
does not recognize both of these features. Actually both are working. Lets see what happens when I get to the testing part.
Is it possible that no MongoDB is the problem? I dont really think so, but if it doesnt resolve until then,
I will have to really use MongoDB. But for the moment, just go on.