feat: Allow delete requests
If an id is provided, the object is deleted and a success message is returned.
If no id is provided, an error message is returned.
If any other type of error occurs, another error message is returned.
Fucking FCC testing does not accept this either.
Lets see after the tests.