feat: FCC POST and GET tests all pass
I changed something important. The problem were the optional fields status_text and assigned_to.
When I returned the object on a POST request i just returned both properties, wether they
were submitted or not. In the browser, it is displayed as an empty string.
Turns out, when console.loggin the return object, it is not an empty string, but undefined.
That was the thing that FCC test suite did  not want to see.